==> analysis/genomic_bins/2_compute_mappablegenome_overlap_run_qmap.ipynb <==
# Compute the overlap of a genomic bin with the mappable genome

Get genomic bins of a given length (1 Mbp) and compute their overlap with the mappbale genome. 

The script runs per chromosome to speed up the analysis. 

The output for a set of bins of N size (1 Mbp) is 24 files (one per chromosome) with coordinates and the size of the overlap per bin in the chromosome.

import pandas as pd

main_dir = ''

input_dir = output_dir = f'{main_dir}/genomic_bins/data'

mappable_genome_file = f'{main_dir}/data/hg38_mappable_genome.nodrivers.tsv.gz'

map_file = f'{main_dir}/code/compute_mappablegenome_overlap_nodrivers.map'
code_file = f'{main_dir}/code/compute_mappablegenome_overlap_as_bed.py'

info = [
    '[params]',
    'cores=1',
    'memory=16G\n',
    '[pre]',
    '. "/home/$USER/miniconda3/etc/profile.d/conda.sh"',
    'conda activate hotspots_framework\n',
    '[jobs]',
]

genome = 'hg38'

# Select length of bins to be computed (in bp)
bin_sizes = [1000000]

chromosomes = list(map(lambda x: f'chr{x}', list(range(1,23)) + ['X', 'Y']))

with open(map_file, 'w') as ofd: 
    for line in info: 
        ofd.write(f'{line}\n')
        
    for chromosome in chromosomes:
        for bin_size in bin_sizes: 
            bins_f = f'{input_dir}/{genome}_{round(bin_size/1e3)}kb_bin.bed.gz'
            output_f = f'{output_dir}/{genome}_{round(bin_size/1e3)}kb_bin.nodrivers.{chromosome}.bed.stats'
            ofd.write(
                f'python {code_file} --chromosome {chromosome} --bins_f {bins_f} --mappable_genome_f {mappable_genome_file} --output_f {output_f}\n')
        print(chromosome)
