==> analysis/genomic_bins/6_mutations_per_bin_run_qmap.ipynb <==
# Mutations per bin

This code intersects filtered mutations in mappable bins with sets of bins of different length (autosomes; driver regions removed)

Mutations in a cancer type are assigned to a signature by maximum likelihood 

Once mutations per signature are intersected with each set of bins, a normalised mutation rate is obtained for the signature across bins

import json

import pandas as pd

cancer_types = [
    'BRCA', 
    'BLADDER_URI', 
    'COADREAD', 
    'ESOPHA_STOMACH', 
    'NSCLC', 
    'PROSTATE', 
    'SKCM'
]

bins = [1000000, 500000, 250000, 100000, 50000, 25000, 10000]

main_dir = ''

input_dir = f'{main_dir}/propensity/data/filtered_mutations'
output_dir = f'{main_dir}/genomic_bins/data/mutations_per_bin'

### Run intersect of filtered mutations (within mappable megabases) and defined bins 

code_f = f'{main_dir}/genomic_bins/code/mutations_per_bin.py'
map_f = f'{main_dir}/genomic_bins/code/mutations_per_bin.map'

info = [
    '[params]',
    'cores=1',
    'memory=16G\n',
    '[pre]',
    '. "/home/$USER/miniconda3/etc/profile.d/conda.sh"',
    'conda activate hotspots_framework\n',
    '[jobs]',
]

with open(map_f, 'w') as ofd: 
    for line in info: 
        ofd.write(f'{line}\n')
    for ctype in cancer_types: 
        input_f = f'{input_dir}/{ctype}_SBS96_total_maxprob.originalref.hg38_1000kb_autosomes_bin.tsv'
        for bin_size in bins: 
            bins_f = f'{main_dir}/genomic_bins/data/hg38_{int(bin_size/1000)}kb_bin.nodrivers.filtered.mappable_positions.autosomes.binids.txt'
            output_f = f'{output_dir}/{ctype}.{int(bin_size/1000)}kb.nodrivers.total_maxprob.mutations_per_bin.json'
            ofd.write(f'python {code_f} -m {input_f} -b {bins_f} -o {output_f}\n')            

### Check

Original number of available bins per size; if some bins are missing in the analysis below it means that either they do not have overlapping mutations or none of their sequence overlaps mappable and driver-free regions
- 1 Mb: 2196
- 500kb: 4392
- 250kb: 8784
- 100kb: 21960
- 50kb: 43920
- 25kb: 87840
- 10kb: 219600

for ctype in cancer_types: 
    print(f'\n{ctype}')
    for bin_size in bins: 
        bin_size = f'{int(bin_size/1000)}kb'
        f = f'{output_dir}/{ctype}.{bin_size}.nodrivers.total_maxprob.mutations_per_bin.json'
        with open(f, 'rt') as fd: 
            data = json.load(fd)
        print(bin_size, len(data.keys()))

### Normalise mutation rate across bins

data_to_run = {
 'BLADDER_URI': ['SBS2', 'SBS13', 'SBS5', 'SBS1'],
 'BRCA': ['SBS5', 'SBS1', 'SBS2', 'SBS13', 'SBS3'],
 'COADREAD': ['SBS1', 'SBS40', 'SBS17b', 'SBS5', 'SBS18', 'SBS93', 'SBS17a'],
 'ESOPHA_STOMACH': ['SBS17b', 'SBS17a', 'SBS5', 'SBS1'],
 'NSCLC': ['SBS4', 'SBS5', 'SBS40', 'SBS2', 'SBS13', 'SBS1'],
 'PROSTATE': ['SBS5', 'SBS1', 'SBS8'],
 'SKCM': ['SBS7a', 'SBS7b']
}

for bin_size in bins:
    pseudocount = int(bin_size)/1000000
    print(pseudocount)

# Load mutations per bin

for bin_size in bins:
    
    pseudocount = int(bin_size)/1000000
    
    bin_size_str = f'{int(bin_size/1000)}kb'
    # get bins available
    bins_f = f'{main_dir}/genomic_bins/data/hg38_{bin_size_str}_bin.nodrivers.filtered.mappable_positions.autosomes.binids.txt'
    bins_df = pd.read_csv(bins_f, sep='\t', header=0)
    total_bins = set(bins_df['BINID'].unique())
    
    for ctype, signatures in data_to_run.items(): 
        
        # Read mutations per Mb
        muts_f = f'{output_dir}/{ctype}.{bin_size_str}.nodrivers.total_maxprob.mutations_per_bin.json'
        with open(muts_f, 'r') as fd: 
            m_per_bin = json.load(fd)

        for signature in signatures:

            # Get mutations 
            mutations = dict()
            for binid in total_bins: 
                
                mbinid_data = m_per_bin.get(binid)
                
                # If bin has at least 1 mutation from any signature
                if mbinid_data: 
                    mutations[binid] = mbinid_data.get(signature, 0) + pseudocount
                else: 
                    mutations[binid] = pseudocount
                    
            total_mutations = sum(mutations.values())
            muts_norm = dict([(binid, counts/total_mutations) for binid, counts in mutations.items()])

            output_f = f'{output_dir}/{ctype}_{signature}.{bin_size_str}.nodrivers.normmuts.total_maxprob.mutations_per_bin.relat_pcount.json'
            with open(output_f, 'w') as ofd: 
                json.dump(muts_norm, ofd) 

            print(bin_size, ctype, signature, len(muts_norm.keys()), sum(muts_norm.values()))
