==> analysis/genomic_bins/8a_liftover_bins.ipynb <==
# Liftover hg38 bins to hg19

Data from large-scale covariates is in hg19. We will liftover 1 Mbp coordinate bins from hg38 to hg19

from pyliftover import LiftOver
lo = LiftOver('hg38', 'hg19')

main_dir = ''

for bin_size in [1000000]: 
    bins_f = f'{main_dir}/genomic_bins/data/hg38_{int(bin_size/1000)}kb_bin.nodrivers.filtered.mappable_positions.autosomes.binids.txt'
    bins_out_f = f'{main_dir}/genomic_bins/data/hg38_{int(bin_size/1000)}kb_bin_liftover_hg19.nodrivers.filtered.mappable_positions.autosomes.binids.txt'
    
    with open(bins_out_f, 'w') as ofd: 
        ofd.write('{}\n'.format('\t'.join(['CHR', 'START', 'END', 'BINID'])))
        with open(bins_f, 'r') as fd: 
            next(fd)
            for line in fd: 
                binid = line.strip()
                chrom, start_end = binid.split(':')
                start, end = start_end.split('-')
                hg19_start = lo.convert_coordinate(chrom, int(start))
                hg19_end = lo.convert_coordinate(chrom, int(end))

                if not hg19_start or not hg19_end: 
                    print(binid)
                else: 
                    ofd.write('{}\n'.format('\t'.join([chrom, str(hg19_start[0][1]), str(hg19_end[0][1]), binid])))
