==> analysis/genomic_bins/8b_generate_bins_covariates_run_qmap.ipynb <==
# Annotate 1 Mbp bins with large scale covariates

This notebook runs different scripts to intersect bin coordinates with large-scale covariates, including: 
    
- Cromatin accessibility (DNase)
- Replication timming (Repli-seq)
- Gene expression (RNA-seq)

import os

import pandas as pd

info = [
    '[params]',
    'cores=1',
    'memory=4G\n',
    '[pre]',
    '. "/home/$USER/miniconda3/etc/profile.d/conda.sh"',
    'conda activate pybigwig\n',
    '[jobs]',
]

main_dir = ''

bins_f = f'{main_dir}/genomic_bins/data/hg38_1000kb_bin_liftover_hg19.nodrivers.filtered.mappable_positions.autosomes.binids.txt'

### DNase

map_file = f'{main_dir}/genomic_bins/code/dnase_ctype.map'
code_path = f'{main_dir}/genomic_bins/code/dnase_ctype.py'

metadata_f = f'{main_dir}/genomic_bins/code/Roadmap.metadata.consolidated.txt.csv'
cov_f = f'{main_dir}/genomic_bins/code/covariates_list.csv'

url='https://egg2.wustl.edu/roadmap/data/byFileType/signal/consolidated/macs2signal/foldChange/'

output_path = f'{main_dir}/genomic_bins/data/large_scale_cov/dnase'

histone_marks = ['DNase']

cancertypes_df = pd.read_csv(cov_f, sep='\t', header=0)
cancertypes = cancertypes_df['CANCER_TYPE'].tolist()
cancertypes

with open(map_file, 'w') as ofd: 
    for line in info: 
        ofd.write(f'{line}\n')
    for ctype in cancertypes: 
        for mark in histone_marks: 
            output_f = f'{output_path}/{ctype}_hg38_1000kb_bin.{mark}.filtered.bed.gz'
            ofd.write(f'python {code_path} --cancer-type {ctype} --bins_f {bins_f} --metadata_f {metadata_f} --output_f {output_f} --cov_f {cov_f} --url {url}\n')

map_file

# RNA

map_file = f'{main_dir}/genomic_bins/code/rna_ctype.map'
code_path = f'{main_dir}/genomic_bins/code/rna_ctype.py'

metadata_f = f'{main_dir}/genomic_bins/code/epigenome_names.csv'
cov_f = f'{main_dir}/genomic_bins/code/covariates_list.csv'

url = 'https://egg2.wustl.edu/roadmap/data/byDataType/rna/signal/normalized_bigwig/stranded/'

output_path = f'{main_dir}/genomic_bins/data/large_scale_cov/rna'

cancertypes_df = pd.read_csv(cov_f, sep='\t', header=0)
cancertypes = cancertypes_df['CANCER_TYPE'].tolist()
cancertypes

bins = [1000000]

with open(map_file, 'w') as ofd: 
    for line in info: 
        ofd.write(f'{line}\n')
    for ctype in cancertypes: 
        for binsize in bins: 
            output_f = f'{output_path}/{ctype}_hg38_1000kb_bin.rna.filtered.bed.gz'
            ofd.write(f'python {code_path} --cancer-type {ctype} --bins_f {bins_f} --metadata_f {metadata_f} --output_f {output_f} --cov_f {cov_f} --url {url}\n')

# Replication timing

map_file = f'{main_dir}/genomic_bins/code/replication_ctype.map'
code_path = f'{main_dir}/genomic_bins/code/replication_ctype.py'

output_path = f'{main_dir}/genomic_bins/data/large_scale_cov/replication'

url = 'http://hgdownload.cse.ucsc.edu/goldenPath/hg19/encodeDCC/wgEncodeUwRepliSeq/'

bins = [1000000]
cancertypes = [
    'SOLID',
    'NON_SOLID'
]

with open(map_file, 'w') as ofd: 
    for line in info: 
        ofd.write(f'{line}\n')
    for ctype in cancertypes: 
        for binsize in bins: 
            output_f = f'{output_path}/{ctype}_hg38_1000kb_bin.RepliSeq.filtered.bed.gz'
            ofd.write(f'python {code_path} --cancer-type {ctype} --bins_f {bins_f} --output_f {output_f} --url {url}\n')
