==> analysis/genomic_bins/8c_merge_largescale_run_qmap.ipynb <==
# Annotate 1 Mbp bins with mutations inside and outside hotspots

Annotate each 1 Mbp bin with mutations inside and outside hotspots considering their signature mutational probabilities (vector)

Note that some paths are hardcoded in the Python script "merge_largescale.py" and need to be updated

import pandas as pd

main_dir = ''

mutations_dir = f'{main_dir}/signatures/assign_muts_to_sigs/data'

output_dir = f'{main_dir}/genomic_bins/data/large_scale_cov/mutations'

bins_f = f'{main_dir}/genomic_bins/data/hg38_1000kb_bin.nodrivers.filtered.mappable_positions.autosomes.binids.txt'

map_file = f'{main_dir}/genomic_bins/code/mutations_per_bin_vector.map'
code_file = f'{main_dir}/genomic_bins/code/mutations_per_bin_vector.py'

covariates_file = f'{main_dir}/genomic_bins/code/covariates_list.csv'
cancertypes_df = pd.read_csv(covariates_file, sep='\t', header=0)
cancertypes = cancertypes_df['CANCER_TYPE'].tolist()
cancertypes

info = [
    '[params]',
    'cores=1',
    'memory=8G\n',
    '[pre]',
    '. "/home/$USER/miniconda3/etc/profile.d/conda.sh"',
    'conda activate hotspots_framework\n',
    '[jobs]',
]

with open(map_file, 'w') as ofd: 
    for line in info: 
        ofd.write(f'{line}\n')
    for ctype in cancertypes: 
        for mtype in ['in', 'out']: 
            input_f = f'{mutations_dir}/{ctype}_SBS96_{mtype}.txt'
            output_f = f'{output_dir}/{ctype}_SBS96_{mtype}.mutation_vector_sigs.1000kb_bins.json'
            ofd.write(f'python {code_file} -m {input_f} -b {bins_f} -o {output_f}\n')
